切り出した関数をまとめる
